Divide and organize classes into packages
