+++ build.gradle
@@ -28,7 +28,7 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.Duke"
+    mainClassName = "duke.Duke.Duke"
 }
 
 shadowJar {
