+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,19 +1,25 @@
+package duke.commands;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+import duke.task.TaskList;
+import duke.ui.Ui;
+import duke.storage.Storage;
+
+public class MarkCommand extends Command {
     private final int index;
 
-    UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.unmark(this.index);
+        tasks.mark(this.index);
         int taskNum = index + 1;
         try {
             storage.writeToFile(tasks);
-            ui.showToUser("You have marked task " + taskNum + " as undone.");
+            ui.showToUser("You have marked task " + taskNum + " as done.");
         } catch (IOException e) {
             ui.showError("Unable to write to file. Please run Duke again.");
         }
